Test under /tmp rather than under ./
For most Perl libraries, the correct way to create a directory for the purpose
of testing is to use File::Temp.  File::Path cannot use that as File::Temp has
a dependency on File::Path.  But what do we do if we do not want to run tests
under the current directory -- and particularly not if the user is a
CPANtester?

The next best alternative might be to use File::Spec::Functions::tmpdir() to
identify the system's own preference for a tempdir, chdir to that directory,
then -- as the tests in t/Path.t currently do -- compose names for testing
directories and files using other File:Spec::Functions functions and use them
during testing, manually cleaning up everything at the end.

This branch explores that approach -- what might be called a change-to-tempdir
approach to testing. t/via_tmpdir.t takes this approach.  Currently it holds
most of the tests found in t/Path.t except (a) those that currently need to be
run by root; and (b) those that rely on Test::Output to match against STDOUT
and STDERR.

This branch is, at this point, purely a thought experiment and is not intended
to be the basis for a patch of pull request.