This branch is a failure.
I had hoped to separate the initial processing of the hash of options passed
to make_path or remove_tree, respectively, from the assignments to the $arg
hash passed internally to _mkpath and _rmtree.

However, I was confounded by the weird error handling setup.  If the user
wants to track/trap errors, she must pass:

  { error => \$error }

Internally, the value of the 'error' key is a list of errors, each of which is
itself a key-value pair (hashref).  So far, so good.  But where in memory is
this list being held?  In the caller program!  So that means that inside, say,
make_path -> mkpath -> _mkpath, $arg->error (as it exists in master) or
$options->error (as I attempted to revise it in this branch) is always
pointing *outside* the scope of package File::Path.

If this were object-oriented, we'd accumulate the list of non-fatal errors in
an attribute of the object.  We'd provide the user with a method to get that
list.  But this package is not OO.

Could we store the list of errors in a package-scoped lexical variable?  We'd
have to say 'use vars' to maintain File::Path's compatibility with pre-5.6
perls.  But even then the current interface handling would have to go.